move qword to submod
